Fixed another critical bug, Pushed persistence
Made it so that the setAlarm widget now shows up with the time that you had previously set on it. I also fixed the default difficulty setting. Basically the same thing that was happening to the stop alarm activity was happening to the main activity, so I fixed that as well.