File Changes
I've gone ahead and uploaded everything required to make this project and run it on your own system. This is now a little more user friendly. If anyone wants to download this repository and work with it, they can without having to guess on a lot of the crucial build configs. Enjoy!